requirements.txt
fastapi
uvicorn
openai
httpx
asyncio
pydantic
aiohttp
python-dotenv
aiohttp